+++ Capstone_stage2/app/src/main/java/com/example/riteshkumarsingh/capstone_stage2/data/source/remote/RemoteDataSource.java
@@ -53,12 +53,12 @@ public class RemoteDataSource implements DataSource{
     }
 
     @Override
-    public Observable<MovieDetails> getMovieDetails(int movie_id) {
+    public Observable<MovieDetails> getMovieDetails(Long movie_id) {
         return mApiService.getMovieDetails(movie_id);
     }
 
     @Override
-    public Observable<MovieVideos> getMovieVideos(int movie_id) {
+    public Observable<MovieVideos> getMovieVideos(Long movie_id) {
         return mApiService.getMovieVideos(movie_id);
     }
 }
